Dram pmem benchmark (#95)
* Initial hybrid memory implementation.

* Fix NUMA bug.

* Added handling of file descriptor and unmapping of dram-mapped file.

* Removed config variable memory type. Removed pmdk from dependencies.

* Revert CMakeList.

* Add app dram and path parsing.

* Run formatter.

* Refactored dram pmem execution.

* Weird fix.

* Add flags.

* Fix dram flags.

* Run formatter after merge.

* Remove fd from functions.

* Apply suggestion from PR.

* Formatter.

* Apply more suggestions from PR.

* Add parentheses.

* Refactor is dram variable.

* Move flags to utils.hpp.

* Do not share DRAM data for now.

* Fix tests.

* Fix flags for tests.

* Add single run tests for DRAM.

* Fixes after merge.

* Remove imports.

* Refactor parallel run.

* Remove import.

* Remove more imports.

* Add todo.

+++ src/single_benchmark.cpp
@@ -1,7 +1,6 @@
 #include "single_benchmark.hpp"
 
 #include <csignal>
-#include <utility>
 
 namespace {
 
@@ -43,10 +42,10 @@ void SingleBenchmark::set_up() {
 
 void SingleBenchmark::tear_down(const bool force) {
   if (!pmem_data_.empty() && pmem_data_[0] != nullptr) {
-    pmem_unmap(pmem_data_[0], configs_[0].total_memory_range);
+    munmap(pmem_data_[0], configs_[0].total_memory_range);
     pmem_data_[0] = nullptr;
   }
-  if (!owns_pmem_files_.empty() && owns_pmem_files_[0] || force) {
+  if (configs_[0].is_pmem && (!owns_pmem_files_.empty() && (owns_pmem_files_[0] || force))) {
     std::filesystem::remove(pmem_files_[0]);
   }
 }
