+++ src/utils.cpp
@@ -1,15 +1,16 @@
 #include "utils.hpp"
 
+#include <fcntl.h>
+#include <spdlog/spdlog.h>
+#include <unistd.h>
+
 #include <algorithm>
 #include <fstream>
 #include <random>
 #include <thread>
-#include <vector>
 
 #include "json.hpp"
-#include "libpmem.h"
 #include "read_write_ops.hpp"
-#include "spdlog/spdlog.h"
 
 #ifdef HAS_NUMA
 #include <numa.h>
@@ -18,49 +19,45 @@
 
 namespace perma {
 
-char* map_pmem_file(const std::filesystem::path& file, const size_t expected_length) {
-  int is_pmem;
-  size_t mapped_length;
-  void* pmem_addr = pmem_map_file(file.c_str(), 0, 0, 0, &mapped_length, &is_pmem);
-  if (pmem_addr == nullptr || (unsigned long)pmem_addr == 0xFFFFFFFFFFFFFFFF) {
+void internal::setPMEM_MAP_FLAGS(const int flags) { internal::PMEM_MAP_FLAGS = flags; }
+
+char* map_file(const std::filesystem::path& file, const bool is_dram, size_t expected_length) {
+  uint64_t fd = -1;
+  int flags;
+  if (!is_dram) {
+    const mode_t mode = 0644;
+    fd = open(file.c_str(), O_RDWR | O_DIRECT, mode);
+    if (fd == -1) {
+      throw std::runtime_error{"Could not open file: " + file.string()};
+    }
+    flags = internal::PMEM_MAP_FLAGS;
+  } else {
+    flags = internal::DRAM_MAP_FLAGS;
+  }
+
+  void* addr = mmap(nullptr, expected_length, PROT_READ | PROT_WRITE, flags, fd, 0);
+  close(fd);
+  if (addr == MAP_FAILED) {
     throw std::runtime_error{"Could not map file: " + file.string()};
   }
 
-  if (mapped_length != expected_length) {
-    throw std::runtime_error("Existing pmem data file has wrong size.");
-  }
-
-  if (!is_pmem) {
-    spdlog::warn("File {} is not in persistent memory!", file.string());
-  }
-
-  return static_cast<char*>(pmem_addr);
+  return static_cast<char*>(addr);
 }
 
-char* create_pmem_file(const std::filesystem::path& file, const size_t length) {
-  const std::filesystem::path base_dir = file.parent_path();
-  if (!std::filesystem::exists(base_dir)) {
-    if (!std::filesystem::create_directories(base_dir)) {
-      throw std::runtime_error{"Could not create dir: " + base_dir.string()};
+char* create_file(const std::filesystem::path& file, const bool is_dram, size_t length) {
+  if (!is_dram) {
+    const std::filesystem::path base_dir = file.parent_path();
+    if (!std::filesystem::exists(base_dir)) {
+      if (!std::filesystem::create_directories(base_dir)) {
+        throw std::runtime_error{"Could not create dir: " + base_dir.string()};
+      }
     }
-  }
 
-  int is_pmem;
-  size_t mapped_length;
-  void* pmem_addr = pmem_map_file(file.c_str(), length, PMEM_FILE_CREATE, 0644, &mapped_length, &is_pmem);
-  if (pmem_addr == nullptr || (unsigned long)pmem_addr == 0xFFFFFFFFFFFFFFFF) {
-    throw std::runtime_error{"Could not create file: " + file.string() + " (error: " + std::strerror(errno) + ")"};
+    std::ofstream temp_stream{file};
+    temp_stream.close();
+    std::filesystem::resize_file(file, length);
   }
-
-  if (!is_pmem) {
-    spdlog::warn("File {} is not in persistent memory!", file.string());
-  }
-
-  if (length != mapped_length) {
-    throw std::runtime_error{"Mapped size different than specified size!"};
-  }
-
-  return static_cast<char*>(pmem_addr);
+  return map_file(file, is_dram, length);
 }
 
 std::filesystem::path generate_random_file_name(const std::filesystem::path& base_dir) {
@@ -200,7 +197,8 @@ void log_numa_nodes(const std::vector<uint64_t>& nodes) {
   spdlog::info("Setting NUMA-affinity to node{}: {}", nodes.size() > 1 ? "s" : "", used_nodes_str);
 }
 
-std::vector<uint64_t> auto_detect_numa(const std::filesystem::path& pmem_dir, const size_t num_numa_nodes) {
+std::vector<uint64_t> auto_detect_numa(const std::filesystem::path& pmem_dir, const size_t num_numa_nodes,
+                                       const bool is_dram) {
 #ifndef HAS_NUMA
   spdlog::critical("Cannot detect numa nodes without NUMA support.");
   crash_exit();
@@ -208,13 +206,15 @@ std::vector<uint64_t> auto_detect_numa(const std::filesystem::path& pmem_dir, co
   const std::filesystem::path temp_file = generate_random_file_name(pmem_dir);
   // Create random 2 MiB file
   const size_t temp_size = 2u * (1024u * 1024u);
-  char* pmem_data = create_pmem_file(temp_file, temp_size);
+  char* pmem_data = create_file(temp_file, is_dram, temp_size);
   rw_ops::write_data(pmem_data, pmem_data + temp_size);
 
   int numa_node = -1;
   get_mempolicy(&numa_node, NULL, 0, (void*)pmem_data, MPOL_F_NODE | MPOL_F_ADDR);
-  pmem_unmap(pmem_data, temp_size);
-  std::filesystem::remove(temp_file);
+  munmap(pmem_data, temp_size);
+  if (!is_dram) {
+    std::filesystem::remove(temp_file);
+  }
 
   if (numa_node < 0 || numa_node > num_numa_nodes) {
     spdlog::warn("Could not determine NUMA node. Running without NUMA-awareness.");
@@ -261,7 +261,8 @@ void set_numa_nodes(const std::vector<uint64_t>& nodes, const size_t num_numa_no
 #endif
 }
 
-void init_numa(const std::filesystem::path& pmem_dir, const std::vector<uint64_t>& arg_nodes) {
+void init_numa(const std::filesystem::path& pmem_dir, const std::vector<uint64_t>& arg_nodes,
+               const bool is_dram) {  // TODO: handle numa for dram
 #ifndef HAS_NUMA
   if (!arg_nodes.empty()) {
     spdlog::critical("Cannot explicitly set numa nodes without NUMA support.");
@@ -296,7 +297,7 @@ void init_numa(const std::filesystem::path& pmem_dir, const std::vector<uint64_t
     return;
   }
 
-  const std::vector<uint64_t> detected_nodes = auto_detect_numa(pmem_dir, num_numa_nodes);
+  const std::vector<uint64_t> detected_nodes = auto_detect_numa(pmem_dir, num_numa_nodes, is_dram);
   log_numa_nodes(detected_nodes);
   return set_numa_nodes(detected_nodes, num_numa_nodes);
 #endif
